Conditional filter function

Apply a filter if condition is met. This function is a simple wrapper for filter

conditional_filter(df1, cond, varName, varCond)

Arguments

df1
A dataframe

cond
A logical value indicating if filter should be applied

varName
Name of variable to filter on

varCond
A single value for equality filter or a vector for is member filter

See also

Examples

data(starwars, package = "dplyr")
starwars %>%
  conditional_filter(TRUE, "species", "Human")

#> # A tibble: 35 x 13
#>    name  height  mass hair_color skin_color eye_color birth_year gender
#>    <chr>  <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr> 
#>  1 Luke~    172    77 blond      fair       blue            19   male  
#>  2 Dart~    202   136 none       white      yellow          41.9 male  
#>  3 Leia~    150    49 brown      light      brown           19   female
#>  4 Owen~    178   120 brown, gr~ light      blue            52   male  
#>  5 Beru~    165    75 brown      light      blue            47   female
#>  6 Bigg~    183    84 black      light      brown           24   male  
#>  7 Obi-~    182    77 auburn, w~ fair       blue-gray       57   male  
#>  8 Anak~    188    84 blond      fair       blue            41.9 male  
#>  9 Wilh~    180    NA auburn, g~ fair       blue            64   male  
#> 10 Han ~    180    80 brown      fair       brown           29   male  
#> # ... with 25 more rows, and 5 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>

starwars %>%
  conditional_filter(TRUE, "species", c("Human", "Droid"))

#> # A tibble: 40 x 13
#>    name  height  mass hair_color skin_color eye_color birth_year gender
#>    <chr>  <int> <dbl> <chr>      <chr>      <chr>          <dbl> <chr> 
#>  1 Luke~    172    77 blond      fair       blue            19   male  
#>  2 C-3PO    167    75 NA         gold       yellow         112   NA    
#>  3 R2-D2     96    32 NA         white, bl~ red             33   NA    
#>  4 Dart~    202   136 none       white      yellow          41.9 male  
#>  5 Leia~    150    49 brown      light      brown           19   female
#>  6 Owen~    178   120 brown, gr~ light      blue            52   male  
#>  7 Beru~    165    75 brown      light      blue            47   female
#>  8 R5-D4     97    32 NA         white, red red             NA   NA    
#>  9 Bigg~    183    84 black      light      brown           24   male  
#> 10 Obi-~    182    77 auburn, w~ fair       blue-gray       57   male  
#> # ... with 30 more rows, and 5 more variables: homeworld <chr>, species <chr>,
#> #   films <list>, vehicles <list>, starships <list>