Pauseable reactive

Create a pauseable reactive. Code from jcheng5 https://gist.github.com/jcheng5/cc6647a4971b8125c1bb

pauseableReactive(x, env = parent.frame(), quoted = FALSE,
  priority = 0, domain = shiny::getDefaultReactiveDomain(),
  paused = TRUE)

Arguments

x
For reactive, an expression (quoted or unquoted). For is.reactive, an object to test.

env
The parent environment for the reactive expression. By default, this is the calling environment, the same as when defining an ordinary non-reactive expression.

quoted
Is the expression quoted? By default, this is FALSE. This is useful when you want to use an expression that is stored in a variable; to do so, it must be quoted with quote().

domain
See domains.

paused
Indicates the starting state of the pauseable reactive.